web_app/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3 